refactor(viewer): modify prop naming element to annotation
