+++ libs/insight-viewer/src/Annotation/index.ts
@@ -1,4 +1,4 @@
-export { AnnotationOverlay } from './AnnotationOverlay'
+export { AnnotationOverlay } from './components/AnnotationOverlay/AnnotationOverlay'
 export { useAnnotation } from '../Annotation/hooks/useAnnotation'
 
 export type { AnnotationMode } from '../Annotation/types'
